+++ src/nvstatistics/chapter_frame.py
@@ -33,7 +33,10 @@ class ChapterFrame(StatisticsFrame):
         x0 = self._LBL_WIDTH + self._LBL_DIST
         x3 = xMax - self._RIGHT_MARGIN
         xSpan = x3 - x0
-        wcNorm = xSpan / wordsTotal
+        try:
+            wcNorm = xSpan / wordsTotal
+        except ZeroDivisionError:
+            wcNorm = 0
 
         barColor = self.prefs['color_chapter']
         y = self._LBL_HEIGHT