+++ src/nvstatistics/plotline_frame.py
@@ -34,7 +34,10 @@ class PlotlineFrame(StatisticsFrame):
         x0 = self._LBL_WIDTH + self._LBL_DIST
         x3 = xMax - self._RIGHT_MARGIN
         xSpan = x3 - x0
-        wcNorm = xSpan / wordsTotal
+        try:
+            wcNorm = xSpan / wordsTotal
+        except ZeroDivisionError:
+            wcNorm = 0
 
         barColor = self.prefs['color_plotline']
         y = self._LBL_HEIGHT